Split out domain commands into own action
Splits out the various domain commands into their own action. Long-term
we will use an ActionChain to join everything up together instead.
